Refactor. WIP
