+++ app/models/search/getEntriesByBasicTextSearch.ts
@@ -1,0 +1,32 @@
+import { DEFAULT_PAGE_SIZE } from "../entry.server"
+import { Prisma } from "@prisma/client"
+import { prisma } from "~/db.server"
+import { SEARCH_WILDCARD } from "../search.server"
+import type { Entry } from "../entry.server"
+import type { SearchResultParams } from "../search.server"
+
+export function getEntriesByBasicTextSearch({
+  searchTerm,
+  skip = 0,
+  take = DEFAULT_PAGE_SIZE,
+  caseSensitive = false,
+  database,
+  isUserAdmin = false,
+}: SearchResultParams) {
+  const searchWildcard =
+    searchTerm === SEARCH_WILDCARD ? "%" : `%${searchTerm}%`
+
+  return prisma.$queryRaw<Pick<Entry, "id" | "headword" | "is_public">[]>`
+  SELECT
+    id, headword, is_public
+  FROM det_entries
+  WHERE
+    IF(${caseSensitive},
+      (headword) LIKE (${searchWildcard}),
+      LOWER(headword) LIKE LOWER(${searchWildcard})
+    )
+    AND (det_entries.dchp_version IN (${Prisma.join(database)}))
+    AND (det_entries.is_public = 1 OR ${isUserAdmin})
+  ORDER BY headword ASC LIMIT ${take} OFFSET ${skip}
+  `
+}
